Added composer support
