+++ .gitignore
@@ -1,0 +1,3 @@
+.idea/
+vendor/
+composer.phar
