+++ examples/search_streets.php
@@ -1,5 +1,6 @@
 <?
-include '../kladr.php';
+
+require __DIR__ . '/../vendor/autoload.php';
 
 // Инициализация api, в качестве параметров указываем токен и ключ для доступа к сервису
 $api = new Kladr\Api('51dfe5d42fb2b43e3300006e', '86a2c2a06f1b2451a87d05512cc2c3edfdf41969');
