+++ examples/simple.php
@@ -1,5 +1,6 @@
 <?php
-include '../kladr.php';
+
+require __DIR__ . '/../vendor/autoload.php';
 
 $api = new Kladr\Api('51dfe5d42fb2b43e3300006e', '86a2c2a06f1b2451a87d05512cc2c3edfdf41969');
