+++ src/ObjectType.php
@@ -1,0 +1,34 @@
+<?php
+
+namespace Kladr;
+
+/**
+ * Перечисление типов объектов
+ */
+class ObjectType
+{
+    /**
+     * Регион
+     */
+    const Region = 'region';
+
+    /**
+     * Район
+     */
+    const District = 'district';
+
+    /**
+     * Населённый пункт
+     */
+    const City = 'city';
+
+    /**
+     * Улица
+     */
+    const Street = 'street';
+
+    /**
+     * Строение
+     */
+    const Building = 'building';
+}
